media site restructure
